Move classes to respective modules, import classes per module following webpack principles
