+++ src/game.js
@@ -1,3 +1,7 @@
+import Phaser from './phaser';
+import GameMenuScene from '../src/scenes/GameMenuScene';
+import GameScene from '../src/scenes/GameScene';
+import GameOverScene from '../src/scenes/GameOverScene';
 const config = {
 	type: Phaser.AUTO,
 	width: 1080,
